requirements.txt
streamlit
torch
torchvision
numpy
Pillow
matplotlib
seaborn
joblib
scikit-learn
scikit-image
opencv-python